moved the config impl into its own file.
